@@ -3,7 +3,7 @@ package jpabook.jpashop.controller;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotEmpty;
 
 @Getter @Setter
 public class MemberForm {
